add UTs to add interface nics to cniResult
